backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
supabase==2.3.5
supabase-py==2.3.5
cloudinary==1.36.0
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
cryptography==42.0.5
